Separate integration tests for BaremetalMachine Manager
